Implement more effecient sprint creator
Implemented a more effecient sprint creator call the micro sprint
creator that works by more closely representing how I actually
input the data. This should reduce duplication in the input yaml
but make it less flexible overall. For this reason the original sprint
creator has also been left in. The micro sprint creator is untested as of yet.

Pulled the JiraController class into it's own file so it can be shared
by the micro and regular sprint creators.

Removed the identities from the sample yamls.